Fix for tornado 4.0 and above
In tornado 4.0 a same origin check has been set that checks that the
origin header set by the browser is the same as the host header.

For proxied connections we need to override this behavior.

The fix will accept an origin pattern to be matched by the origin header
sent by the browser as an option or it will accept any origin if nothing
is provided.

Examples:

To not check origin use it as it is (default behavior):
python bucket.py

To allow only connections from *.mydomain.com:
python bucket.py --origin=^.*?\.mydomain\.com


Related Links:
* http://stackoverflow.com/questions/24800436/under-tornado-v4-websocket-connections-get-refused-with-403
* http://www.tornadoweb.org/en/stable/releases/v4.0.0.html#backwards-compatibility-notes